+++ waves/_abaqus/_settings.py
@@ -1,3 +1,5 @@
+import typing
+
 # Default supported file extensions
 _default_h5_extension = ".h5"
 _default_yaml_extension = ".yaml"
@@ -5,6 +7,6 @@ _default_parsed_extension = ".parsed"
 _default_odbreport_extension = ".csv"
 _default_timestamp_format = "%Y%m%d-%H%M%S"
 _default_abaqus_command = "abq2024"
-_default_xarray_engine = "h5netcdf"
+_default_xarray_engine: typing.Literal["netcdf4", "scipy", "h5netcdf"] = "h5netcdf"
 # Temporary contact info
 _parsing_error_contact = "prabhu@lanl.gov"